move the new chart and mid tests to the Parsing folder
